+++ Group4SlackBot/src/main/resources/application.properties
@@ -16,3 +16,4 @@ slashQuizToken=k5aba1n75fgWadaScbmkrKb6
 #Web hook URLS
 slackIncomingWebhookUrl=https://hooks.slack.com/services/T025DUwebhookurloOYvPiHL7y6
 
+
