+++ Group4SlackBot/src/main/resources/quiz.json
@@ -1,7 +1,7 @@
 {
   "1":{
     "question": "What is the capital of California?",
-    "a": "Las Angeles",
+    "a": "Los Angeles",
     "b": "Sacramento",
     "c": "Gilroy",
     "d": "San Francisco",
